+++ ngraph/core/src/op/normalize_l2.cpp
@@ -74,7 +74,7 @@ void op::NormalizeL2::pre_validate_and_infer_types()
             for (auto axis : reduction_axes)
             {
                 NODE_VALIDATION_CHECK(this,
-                                      axis < input_rank.get_length(),
+                                      static_cast<int64_t>(axis) < input_rank.get_length(),
                                       "Reduction axis (",
                                       axis,
                                       ") is out of bounds ",
