+++ ngraph/core/src/op/space_to_depth.cpp
@@ -117,7 +117,7 @@ bool ngraph::op::v0::SpaceToDepth::evaluate_space_to_depth(const HostTensorVecto
     const size_t spatial_dim_index = 2;
     const size_t spatial_dims = data_shape.size() - spatial_dim_index;
 
-    for (int i = spatial_dim_index; i < data_shape.size(); ++i)
+    for (size_t i = spatial_dim_index; i < data_shape.size(); ++i)
     {
         NODE_VALIDATION_CHECK(this,
                               m_blocksize > 0 && data_shape.at(i) % m_blocksize == 0,
@@ -133,7 +133,7 @@ bool ngraph::op::v0::SpaceToDepth::evaluate_space_to_depth(const HostTensorVecto
     // rearrange them so as appropriate chunks of data where close to their
     // destination place. Finally squeeze data from respective dimensions.
     Shape dispersed_shape{n_dim, c_dim};
-    for (int i = 0; i < spatial_dims; ++i)
+    for (size_t i = 0; i < spatial_dims; ++i)
     {
         dispersed_shape.push_back(data_shape.at(i + spatial_dim_index) / m_blocksize);
         dispersed_shape.push_back(m_blocksize);
@@ -199,7 +199,7 @@ bool ngraph::op::v0::SpaceToDepth::evaluate_space_to_depth(const HostTensorVecto
                                  elem_size);
 
     Shape squeezed_shape{n_dim};
-    for (int i = 0; i < spatial_dims; ++i)
+    for (size_t i = 0; i < spatial_dims; ++i)
     {
         squeezed_shape.push_back(data_shape.at(spatial_dim_index + i) / m_blocksize);
     }
