+++ ngraph/core/src/op/util/logical_reduction.cpp
@@ -97,7 +97,7 @@ void op::util::LogicalReduction::validate_and_infer_types()
         }
 
         std::vector<Dimension> dims;
-        for (size_t i = 0; i < input_rank.get_length(); i++)
+        for (int64_t i = 0; i < input_rank.get_length(); i++)
         {
             if (reduction_axes.count(i) == 0)
             {
