+++ ngraph/core/src/op/util/logical_reduction_keep_dims.cpp
@@ -71,7 +71,7 @@ void op::util::LogicalReductionKeepDims::validate_and_infer_types()
             }
 
             std::vector<Dimension> dims;
-            for (size_t i = 0; i < input_rank.get_length(); i++)
+            for (int64_t i = 0; i < input_rank.get_length(); i++)
             {
                 if (reduction_axes.count(i) == 0)
                 {
